straighten out hy_query_filter_updates() semantics.
The previous semantics allowed *extending* the already filtered set to
non-installed packages. This was meant as a convenience for the library
clients but makes little sense. IOW, we shouldn't try to hide the fact
that installed 'foo' and 'foo' available in a repository are two different
entitites.

Not to mention it is expensive to do the resolving just to make sure the
newer version is really installable and we already have a goal for that.