change the Yum repo loading API.
the main problem with the old API is that it seems like the client can invoke
the writing/loading functions for main/filelist/presto datatypes in an arbitrary
order while in fact he can not (for instance, things go wrong in libsolv if one
loads the entire repository, then filelists, then write the cache file and then
tries to load everything back from the cache.).

The new approach relies on a single function and flags to specify the
precise operation. Hawkey handles the ordering of the operation. This
seemingly takes away some of the flexibility, like deciding to load
filelists or prestodelta at a later time. Some of these mechanisms might
still be added. Unittests testing specificities of the old API have been
removed for now.

The new API is also extensible with a flag forcing e.g. filelists to only
be loaded on demand.

In practice at this point (current practices of Fedora packaging sometimes
need full filelists for resolving, libsolv being best tuned for loading
everything from cache very fast but filelists load from .xml is very
slow), anything but preparing the cache with all extensions once and then
reusing it as long as possible makes little sense.