Subject: discovering all NEVRA explanations of what user's input meant.
Because of the unfortunate shapes the full package NEVRAS can take (dots,
dashes, numbers and arches in the name etc.), there usually can be found
more than one meaning of what the user meant with his input. This is an
attempt of implementing a fast, standardized way to handle these.

This first implementation is Python only, extensively unit tested. If it
works it will be turned into C implementation with bindings, with the
tests reused.