subject: best shot at 'some-lib-devel' is not that EVR is 'lib-devel'.
This fully shows that the "best guess" parse and "best guess with known names"
parse should not follow the same order of tested parses. But since nobody uses
the simple best guess yet (i.e. no asking the sack if he knows the name), the
patch optimizes for the latter case. The form order will have to split between
the two later.

Resolves: RhBug:903687