Add API to get and set userdata on a HyPackage instance
Resolves: RhBug:977401

We need to do this as typically in users of the API we are passing around arrays
of packages, and we need to be able to add additional data to them. For instance,
we might want to add the location of the .rpm file of the to-be-installed-package
as it's downloaded in librepo, and retrieved before the package is added to an
rpm transaction.

Adding this data means we do not need a package+key metadata dictionary or any
kind of wrapper object around HyPackage when using hawkey in C programs.