Do not enforce all repos load all kinds of specified metadata
Do not return with an error if a repo does not have metadata specified in
hy_sack_load_yum_repo(). We don't want to have to check the repository and find
out if each repo has the desired metadata type. To fix this, turn the meaning
of the flags from 'load this metadata' to 'load this metadata if it exists'.