always allocate memory for reldep
Currently if we try to create reldep from "foo = 1" and foo is
not in sack, it would say that it can't create it. But this is wrong,
because reldep is something abstract and we should check that it doesn't
match anything in query/selector rather than just try to find it immediately.

When we process rich deps, we blindly allocate memory for reldep
because we don't know in advance whether it exists or not.

Libsolv doesn't have any function to validate correctness of name, so we can't
do much (it takes care of parsing rpm-md, so it doesn't check there).

Signed-off-by: Igor Gnatenko <ignatenko@redhat.com>

Closes: #266
Approved by: j-mracek